refactor: Util 패키지 정리
