+++ android/app/src/main/java/com/example/home_rent_app/ui/renthome/viewmodel/RentHomeViewModel.kt
@@ -10,8 +10,8 @@ import com.example.home_rent_app.data.repository.imageurl.ImageUrlRepository
 import com.example.home_rent_app.data.repository.renthome.RentHomeRepository
 import com.example.home_rent_app.data.session.UserSession
 import com.example.home_rent_app.util.FileController
-import com.example.home_rent_app.util.HouseType
-import com.example.home_rent_app.util.RentType
+import com.example.home_rent_app.util.type.HouseType
+import com.example.home_rent_app.util.type.RentType
 import com.example.home_rent_app.util.UiState
 import com.example.home_rent_app.util.logger
 import dagger.hilt.android.lifecycle.HiltViewModel
