+++ android/app/src/main/java/com/example/home_rent_app/ui/searchwanthome/viewmodel/SearchWantHomeViewModel.kt
@@ -11,6 +11,7 @@ import com.example.home_rent_app.data.model.WantHomeResultRequest
 import com.example.home_rent_app.data.repository.wanthomeresult.WantHomeResultRepository
 import com.example.home_rent_app.util.CoroutineException
 import com.example.home_rent_app.util.UiState
+import com.example.home_rent_app.util.logger
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.CoroutineExceptionHandler
 import kotlinx.coroutines.channels.BufferOverflow
@@ -30,7 +31,7 @@ class SearchWantHomeViewModel @Inject constructor(
     ViewModel() {
 
     private val _searchWord = MutableSharedFlow<String>()
-    val searchWord = _searchWord.debounce { 400 }
+    val searchWord = _searchWord.debounce { 400L }
 
     private val _wantHomeResult =
         MutableStateFlow<UiState<PagingData<WantedArticle>>>(UiState.Loading)
@@ -62,6 +63,7 @@ class SearchWantHomeViewModel @Inject constructor(
 
     fun getWantHomeResult(wantHomeResultRequest: WantHomeResultRequest) {
         viewModelScope.launch {
+            logger("로그 viewmodel")
             wantHomeResultRepository.getResult(wantHomeResultRequest)
                 .cachedIn(viewModelScope)
                 .collect { response ->
