--- android/app/src/main/java/com/example/home_rent_app/util/LoginCheck.kt
@@ -1,5 +1,0 @@
-package com.example.home_rent_app.util
-
-enum class LoginCheck {
-    HOME, LOGIN, DEFAULT
-}
