+++ android/app/src/main/java/com/example/home_rent_app/util/bindingadapter/TransferBindingAdapter.kt
@@ -4,6 +4,8 @@ import androidx.databinding.BindingAdapter
 import androidx.databinding.InverseBindingAdapter
 import androidx.databinding.InverseBindingListener
 import com.example.home_rent_app.R
+import com.example.home_rent_app.util.type.HouseType
+import com.example.home_rent_app.util.type.RentType
 import com.google.android.material.chip.ChipGroup
 
 @BindingAdapter("roomTypeFilter")
