+++ android/app/src/main/java/com/example/home_rent_app/util/type/HouseType.kt
@@ -1,4 +1,4 @@
-package com.example.home_rent_app.util
+package com.example.home_rent_app.util.type
 
 enum class HouseType(val value: String) {
     ONE_ROOM("ONEROOM"), TWO_ROOM("TWOROOM"), THREE_ROOM("THREEROOM"), EFFICIENCY("OFFICETEL"), SHARE_HOUSE("SHAREHOUSE")
