+++ android/app/src/main/java/com/example/home_rent_app/util/type/Location.kt
@@ -1,4 +1,4 @@
-package com.example.home_rent_app.util
+package com.example.home_rent_app.util.type
 
 enum class Location(val locationName: String) {
     SEOUL("서울특별시"),
