+++ android/app/src/main/java/com/example/home_rent_app/util/type/RentType.kt
@@ -1,4 +1,4 @@
-package com.example.home_rent_app.util
+package com.example.home_rent_app.util.type
 
 enum class RentType(val value: String) {
     MONTHLY("MONTHLY"), JEONSE("JEONSE")
